pyproject.toml
[project]
name = "mujoco-mcp"
version = "0.1.0"
description = "MCP server for AI-controlled robot pose matching using MuJoCo physics simulation"
readme = "README.md"
authors = [
    { name = "JC", email = "jccalvojackson@gmail.com" }
]
requires-python = ">=3.12"
keywords = ["robotics", "mujoco", "simulation", "mcp", "ai", "pose-control", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "loguru>=0.7.3",
    "mcp[cli]>=1.13.1",
    "mujoco>=3.3.5",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "robot-descriptions>=1.20.0",
]

[project.urls]
"Homepage" = "https://github.com/jccalvojackson/mujoco-mcp"
"Repository" = "https://github.com/jccalvojackson/mujoco-mcp"
"Bug Tracker" = "https://github.com/jccalvojackson/mujoco-mcp/issues"

[project.scripts]
mujoco-mcp = "mujoco_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
demo = [
    "agno>=1.8.1",
    "anthropic>=0.64.0",
    "fastapi>=0.116.1",
    "google-genai>=1.32.0",
    "typer>=0.16.1",
    "weave>=0.52.5",
]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.uv.sources]
agno = { git = "https://github.com/agno-agi/agno", subdirectory = "libs/agno", rev = "v2.0" }
